fix(dx11): harden COM and resource lifetimes

+++ DirectX11/Hunting.cpp
@@ -1375,7 +1375,7 @@ static void HuntNext(char *type, std::set<ItemType> *visited,
 				it = visited->begin();
 				*selectedPos = 0;
 			} else {
-				*selectedPos = std::distance(visited->begin(), it);
+				*selectedPos = static_cast<int>(std::distance(visited->begin(), it));
 			}
 			*selected = *it;
 		} else {
@@ -1504,7 +1504,7 @@ static void HuntPrev(char *type, std::set<ItemType> *visited,
 				*selectedPos = size - 1;
 			} else {
 				--it;
-				*selectedPos = std::distance(visited->begin(), it);
+				*selectedPos = static_cast<int>(std::distance(visited->begin(), it));
 			}
 			*selected = *it;
 		} else {
